Enable cyrus-lists to handle email 'from' recipients that come as an Array
cyrus-lists in production has been broken for the last week because
something changed but I'm not sure what. After investigating, I found
that Net::SMTP.send_message was being called with an array for it's
'from email address' argument.  This was causing a 501 error to be
returned by SendGrid.  I'm not sure why our code was working in the
past because this seems to be the expected behavior of the mail gem.  I
put a hack in to get things working again.